Mover kann mit custom made Sprachblase reden Tiles können kopiert werden
Zeitplan:
11.05 - siehe oben
13    - Kreieren der BaseMaps
15    - Spawn Punkte mit Respawn (Neustart Button)
17    - Intro beenden
19    - Code vereinfachen (Nutzung von Methoden in Methoden (SetMapTile durch Click Methode die TIle (ID) zurückkgibt
21    - Level 2 Item Map - Spiel Level
23    - Neue Spielmechanik Features
25    - Level 3 Item Map - Spiel Level
27    - Level 4 Item Map - Spiel Level
29    - Level 5 Item Map - Spiel Level
Ab Juni - Präsentation + Dokumentation
Vorstellen am 18 ? -- Nochmal nachschauen !

+++ Camera.java
@@ -92,7 +92,7 @@ public class Camera {
         this.clickXOffset = clickXOffset;
     }
 
-    public double getClickYOffset() {
+    public int getClickYOffset() {
         return clickYOffset;
     }
